comit tạm

+++ Services/Services/CapacityService.cs
@@ -2,10 +2,12 @@
 using Database.Interfaces;
 using Domain.DTOs;
 using Domain.Entities;
-using Domain.Identities;
 using Services.Interfaces;
 using Services.Interfaces.Common;
+using Services.Middleware;
+using Services.Transform;
 using Services.Validators;
+using System.Reflection;
 
 namespace Services.Services
 {
@@ -23,6 +25,10 @@ namespace Services.Services
         [RequirePermission("Capacity.View")]
         public async Task<(List<CapacityDto> list, int totalCount, int pageNumber)> GetList(string? pSort = "Id", int? pPageNumber = 1, int? pPageSize = 30, string? pKeyword = "")
         {
+            if (CustomMiddleware.CheckPermission("Capacity.View") == false)
+            {
+                throw new UnauthorizedAccessException(IdentityTransform.ForbiddenException());
+            }
             var result = await _capacityRepo.GetAllAsync(null, pKeyword, pSort, pPageNumber, pPageSize);
 
             var list = _mapper.Map<List<CapacityDto>>(result.list);
@@ -33,6 +39,10 @@ namespace Services.Services
         [RequirePermission("Capacity.View")]
         public async Task<CapacityDto> GetDetail(int pId)
         {
+            if (CustomMiddleware.CheckPermission("Capacity.View") == false)
+            {
+                throw new UnauthorizedAccessException(IdentityTransform.ForbiddenException());
+            }
             var capacity = await _capacityRepo.GetDetailAsync(pId);
 
             var capacityDto = _mapper.Map<CapacityDto>(capacity);
@@ -43,6 +53,10 @@ namespace Services.Services
         [RequirePermission("Capacity.Create")]
         public async Task<bool> Create(CapacityDto pCreate)
         {
+            if (CustomMiddleware.CheckPermission("Capacity.Create") == false)
+            {
+                throw new UnauthorizedAccessException(IdentityTransform.ForbiddenException());
+            }
             CapacityValidator validator = new CapacityValidator(_capacityRepo);
             var validationResult = await validator.ValidateAsync(pCreate);
 
@@ -62,6 +76,10 @@ namespace Services.Services
         [RequirePermission("Capacity.Update")]
         public async Task<bool> Update(CapacityDto pUpdate)
         {
+            if (CustomMiddleware.CheckPermission("Capacity.Update") == false)
+            {
+                throw new UnauthorizedAccessException(IdentityTransform.ForbiddenException());
+            }
             CapacityValidator validator = new CapacityValidator(_capacityRepo, false, pUpdate.Id);
             var validationResult = await validator.ValidateAsync(pUpdate);
 
@@ -81,6 +99,10 @@ namespace Services.Services
         [RequirePermission("Capacity.Delete")]
         public async Task<bool> Delete(int pId)
         {
+            if (CustomMiddleware.CheckPermission("Capacity.Delete") == false)
+            {
+                throw new UnauthorizedAccessException(IdentityTransform.ForbiddenException());
+            }
             var result = await _capacityRepo.DeleteAsync(pId);
 
             return result;
